KAN-17 Add Roles

+++ app/routes/tasks.py
@@ -3,6 +3,7 @@ from sqlalchemy.orm import Session
 from ..database import SessionLocal
 from .. import crud, schemas
 from ..auth import oauth2_scheme, verify_token
+from .enums import Role
 
 router = APIRouter(prefix="/tasks", tags=["Tasks"])
 
@@ -11,15 +12,27 @@ def create(task: schemas.TaskCreate, token: str = Depends(oauth2_scheme)):
     credentials_exception = HTTPException(
         status_code=401,
         detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
+        headers={{"WWW-Authenticate": "Bearer"}},
     )
     token_data = verify_token(token, credentials_exception)
     db: Session = SessionLocal()
+    user = crud.get_user(db, token_data.sub)
+    if user.role not in [Role.admin, Role.manager]:
+        raise HTTPException(403, "Forbidden")
     return crud.create_task(db, task)
 
 @router.get("/")
-def list_tasks():
+def list_tasks(token: str = Depends(oauth2_scheme)):
+    credentials_exception = HTTPException(
+        status_code=401,
+        detail="Could not validate credentials",
+        headers={{"WWW-Authenticate": "Bearer"}},
+    )
+    token_data = verify_token(token, credentials_exception)
     db: Session = SessionLocal()
+    user = crud.get_user(db, token_data.sub)
+    if user.role == Role.developer:
+        return crud.get_tasks(db, user.id)
     return crud.get_tasks(db)
 
 @router.put("/{task_id}")
@@ -27,10 +40,13 @@ def update(task_id: int, task: schemas.TaskUpdate, token: str = Depends(oauth2_s
     credentials_exception = HTTPException(
         status_code=401,
         detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
+        headers={{"WWW-Authenticate": "Bearer"}},
     )
     token_data = verify_token(token, credentials_exception)
     db: Session = SessionLocal()
+    user = crud.get_user(db, token_data.sub)
+    if user.role not in [Role.admin, Role.manager]:
+        raise HTTPException(403, "Forbidden")
     updated = crud.update_task(db, task_id, task)
     if not updated:
         raise HTTPException(404, "Task not found")
@@ -41,11 +57,14 @@ def delete(task_id: int, token: str = Depends(oauth2_scheme)):
     credentials_exception = HTTPException(
         status_code=401,
         detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
+        headers={{"WWW-Authenticate": "Bearer"}},
     )
     token_data = verify_token(token, credentials_exception)
     db: Session = SessionLocal()
+    user = crud.get_user(db, token_data.sub)
+    if user.role not in [Role.admin, Role.manager]:
+        raise HTTPException(403, "Forbidden")
     deleted = crud.delete_task(db, task_id)
     if not deleted:
         raise HTTPException(404, "Task not found")
-    return {"message": "Deleted"}
+    return {{"message": "Deleted"}}
